% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{predictPrice}
\alias{predictPrice}
\title{Predict the price of a house based on some user input}
\usage{
predictPrice(input, data)
}
\arguments{
\item{input}{A collection of all the data gathered from the ui input forms}

\item{data}{Data used to build the kknn model. Must contain the same columns as input.}
}
\value{
The predicted price rounded to the nearest 1000
}
\description{
This function will create a kknn model which aims to predict
the sale price of a house based off of the following characteristings:
TotalBsmtSF, GrLivArea, GarageCars, TotRmsAbvGrd, OverallQual
}
